moves storeMeta to own data-access-layer
