+++ src/requests/bitly/index.ts
@@ -1,5 +1,5 @@
 import { endpoints } from './endpoints'
-import { storeResponseMeta } from '../../lib/storeMeta'
+import { storeResponseMeta } from '../../data-access-layer/storeMeta'
 import fetch from 'node-fetch'
 const ACCESS_TOKEN = process.env.BITLY_ACCESS_TOKEN
 const BASE_URL = 'https://api-ssl.bitly.com/v4/'
@@ -13,26 +13,21 @@ interface Endpoint {
 export default async () => {
   const result = await Promise.all(
     endpoints.map(async (endpoint: Endpoint) => {
-      const res = await fetch(
-        `${BASE_URL}${endpoint.url}`,
-        {
-          method: endpoint.method,
-          headers: {
-            Accept: 'application/json',
-            'Content-Type': 'application/json',
-            Authorization: `Bearer ${ACCESS_TOKEN}`,
-          },
-          body:
-            endpoint.data && endpoint.method !== 'GET'
-              ? JSON.stringify(endpoint.data)
-              : undefined,
-        }
-      )
+      const res = await fetch(`${BASE_URL}${endpoint.url}`, {
+        method: endpoint.method,
+        headers: {
+          Accept: 'application/json',
+          'Content-Type': 'application/json',
+          Authorization: `Bearer ${ACCESS_TOKEN}`,
+        },
+        body:
+          endpoint.data && endpoint.method !== 'GET'
+            ? JSON.stringify(endpoint.data)
+            : undefined,
+      })
       const body = await res.text()
       const headers: any = {}
-      res.headers.forEach(
-        (value, name) => (headers[name] = value)
-      )
+      res.headers.forEach((value, name) => (headers[name] = value))
       return {
         wholeURI: res.url,
         endpoint: endpoint.url,
